requirements.txt
fastapi
uvicorn
opencv-python
onnxruntime
PyYAML
requests